requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
bitsandbytes==0.41.0
braceexpand==0.1.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.5
contourpy==1.3.0
crcmod==1.7
cycler==0.12.1
datasets==3.1.0
decord==0.6.0
deepspeed==0.13.5
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
einops-exts==0.0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss==1.9.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.14.0
fire==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.66.2
h11==0.14.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
jieba==0.42.1
Jinja2==3.1.3
jiter==0.7.0
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
latex2mathml==3.77.0
linkify-it-py==2.0.3
lit==18.1.8
-e git+https://github.com/hiyouga/LLaMA-Factory.git@89ba4513ae5786c396df9653c3bc8dfdc5e716af#egg=llamafactory
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mmcls==0.25.0
mmcv-full==1.6.2
mmsegmentation==0.30.0
model-index==0.1.11
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.9.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
open-flamingo @ git+https://github.com/mlfoundations/open_flamingo.git@655f693fbfa04cd6e9a987d960654624d48917cf
open_clip_torch==2.29.0
openai==1.54.3
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
orjson==3.10.7
oss2==2.17.0
packaging==24.1
pandas==2.2.3
peft==0.12.0
pillow==10.2.0
platformdirs==4.3.6
prettytable==3.11.0
protobuf==3.20.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2023.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.4
rich==13.4.2
rouge-chinese==1.0.3
ruff==0.6.9
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.98
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.38.6
streamlit==1.38.0
streamlit-image-select==0.6.0
svgwrite==1.4.3
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.4.1+cpu
torchvision==0.15.2
tqdm==4.66.5
transformers==4.45.0
triton==2.0.0
trl==0.9.6
typer==0.12.5
typing_extensions==4.12.2
tyro==0.8.11
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.31.0
wandb==0.18.5
watchdog==4.0.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.100
websockets==12.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.1
yarl==1.13.1
zipp==3.20.2
